+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 1.3.0
+- Added compatibility with Hyvä Checkout 1.3.x with full Content Security Policy (CSP) support.
+- Improved PayPal and Apple Pay rendering: terms & required-address validation before payment, and clearer error messaging.
 ## 1.0.0
 - Add Hyvä Checkout initial compatibility with Credit Card hosted fields, Alma (3x and 4x), Apple Pay, Paypal integrated payment methods and from HiPay Enterprise hosted page payment service